pyproject.toml
[tool.poetry]
authors = ["long2ice <long2ice@gmail.com>"]
description = "A fast admin dashboard based on FastAPI and TortoiseORM with tabler ui, inspired by Django admin."
documentation = "https://fastapi-admin.github.io"
homepage = "https://github.com/fastapi-admin/fastapi-admin"
include = [
    "LICENSE",
    "README.md",
    "CHANGELOG.md",
    "fastapi_admin/locales/en_US/LC_MESSAGES/messages.mo",
    "fastapi_admin/locales/zh_CN/LC_MESSAGES/messages.mo"
]
keywords = ["fastapi", "admin", "dashboard", "tortoise-orm"]
license = "Apache-2.0"
name = "fastapi-admin"
packages = [
    { include = "fastapi_admin" },
]
readme = "README.md"
repository = "https://github.com/fastapi-admin/fastapi-admin.git"
version = "1.0.4"

[tool.poetry.dependencies]
Babel = "*"
aiofiles = "*"
aioredis = "*"
bcrypt = "*"
fastapi = "*"
jinja2 = "*"
python = "^3.7"
python-multipart = "*"
tortoise-orm = "*"
uvicorn = { version = "*", extras = ["standard"] }
pendulum = "*"
argon2-cffi = "^21.3.0"

[tool.poetry.dev-dependencies]
# test
pytest = "*"
pytest-asyncio = "*"
pytest-mock = "*"
pytest-xdist = "*"
# lint
black = "*"
flake8 = "*"
isort = "*"
mypy = "*"
pylint = "*"
# example
asyncmy = "*"
asyncpg = "*"
python-dotenv = "*"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
